refactor: add organizationId to experiment to simplify authorization checks

+++ convex/experiments.ts
@@ -1,33 +1,19 @@
 import { getAllOrThrow } from "convex-helpers/server/relationships";
 import { v } from "convex/values";
 
-import { type Id } from "./_generated/dataModel";
-import {
-  mutation,
-  type MutationCtx,
-  query,
-  type QueryCtx,
-} from "./_generated/server";
+import { mutation, query } from "./_generated/server";
 import { getIdentityOrThrow } from "./helpers";
-
-async function getExperimentProject(
-  ctx: MutationCtx | QueryCtx,
-  projectId: Id<"projects">,
-  _identity?: Awaited<ReturnType<typeof getIdentityOrThrow>>,
-) {
-  const identity = _identity ?? (await getIdentityOrThrow(ctx));
-  const project = await ctx.db.get(projectId);
-  if (project?.organizationId !== identity.organization.id)
-    throw new Error("User is not authorized to view this project");
-  return project;
-}
 
 export const listProjectExperiments = query({
   args: {
     projectId: v.id("projects"),
   },
   handler: async (ctx, args) => {
-    await getExperimentProject(ctx, args.projectId);
+    const identity = await getIdentityOrThrow(ctx);
+    const project = await ctx.db.get(args.projectId);
+    if (!project) throw new Error("Project not found");
+    if (project.organizationId !== identity.organization.id)
+      throw new Error("User is not authorized to view this project");
     return await ctx.db
       .query("experiments")
       .withIndex("projectId", (q) => q.eq("projectId", args.projectId))
@@ -35,26 +21,16 @@ export const listProjectExperiments = query({
   },
 });
 
-async function _getExperiment(
-  ctx: MutationCtx | QueryCtx,
-  experimentId: Id<"experiments">,
-) {
-  const identity = await getIdentityOrThrow(ctx);
-  const experiment = await ctx.db.get(experimentId);
-  if (!experiment) throw new Error("Experiment not found");
-  const project = await ctx.db.get(experiment.projectId);
-  if (!project) throw new Error("Project not found");
-  if (project.organizationId !== identity.organization.id)
-    throw new Error("User is not authorized to view this experiment");
-  return experiment;
-}
-
 export const getExperiment = query({
   args: {
     id: v.id("experiments"),
   },
   handler: async (ctx, args) => {
-    const experiment = await _getExperiment(ctx, args.id);
+    const identity = await getIdentityOrThrow(ctx);
+    const experiment = await ctx.db.get(args.id);
+    if (!experiment) throw new Error("Experiment not found");
+    if (experiment.organizationId !== identity.organization.id)
+      throw new Error("User is not authorized to view this experiment");
     const personas = await getAllOrThrow(ctx.db, experiment.personas);
     return { ...experiment, personas };
   },
@@ -68,10 +44,14 @@ export const createExperiment = mutation({
   },
   handler: async (ctx, args) => {
     const identity = await getIdentityOrThrow(ctx);
-    await getExperimentProject(ctx, args.projectId, identity);
+    const project = await ctx.db.get(args.projectId);
+    if (!project) throw new Error("Project not found");
+    if (project.organizationId !== identity.organization.id)
+      throw new Error("User is not authorized to view this project");
 
     return await ctx.db.insert("experiments", {
       name: args.name,
+      organizationId: identity.organization.id,
       // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
       owner: identity.email!,
       personas: args.personas,
@@ -85,7 +65,11 @@ export const deleteExperiment = mutation({
     id: v.id("experiments"),
   },
   handler: async (ctx, args) => {
-    await _getExperiment(ctx, args.id);
+    const identity = await getIdentityOrThrow(ctx);
+    const experiment = await ctx.db.get(args.id);
+    if (!experiment) throw new Error("Experiment not found");
+    if (experiment.organizationId !== identity.organization.id)
+      throw new Error("User is not authorized to delete this experiment");
     await ctx.db.delete(args.id);
   },
 });
@@ -97,7 +81,11 @@ export const editExperiment = mutation({
     personas: v.array(v.id("personas")),
   },
   handler: async (ctx, { id, ...args }) => {
-    await _getExperiment(ctx, id);
+    const identity = await getIdentityOrThrow(ctx);
+    const experiment = await ctx.db.get(id);
+    if (!experiment) throw new Error("Experiment not found");
+    if (experiment.organizationId !== identity.organization.id)
+      throw new Error("User is not authorized to edit this experiment");
     await ctx.db.patch(id, args);
   },
 });
